Fix issue #3 - 'naive' I2C transaction in pimoroniTrackball::changed() that caused it to fail on ESP32 and maybe other targets Fix compiler warnings for use of TwoWire.requestFrom(int, int, int) Fix some style issues and tidy up

+++ src/pimoroniTrackball.h
@@ -64,7 +64,7 @@ class pimoroniTrackball
 		bool release();															//Was the trackball released
 	protected:
 	private:
-		uint8_t _trackballAddress = 0x0A;									//I2C address for the trackball
+		int _trackballAddress = 0x0A;										//I2C address for the trackball
 		TwoWire *_i2cPort = nullptr;										//Pointer to I2C port used by library
 		const uint16_t trackballDeviceId = 0xBA11;							//Device ID for the trackball
 		uint8_t _lastState[5] = {0, 0, 0, 0, 0b10000000};					//Last state of the trackball, which resets on read
